GHUB-0170: close the clang-tidy backlog and turn enforcement on
The remainder of the list, and it now stands at zero. WarningsAsErrors is
on, which the item names as the moment.

Unchecked return values, all nine in main.cpp and all the same call:
fprintf to stderr, whose byte count every program discards. Each says so
with a (void) cast and AllowCastToVoid is set, rather than the check being
excused wholesale -- so a dropped return anywhere it matters is still
reported.

Signed char handed to int, all nine. Every one is a qint8 read off a save
stream -- a suit, a rank, a seat, a row -- range-checked before it is used,
and the check's own remedy would break those guards: casting to unsigned
char first turns a rejected -1 into 255 and walks it past the test above it.
The hazard the check exists for is a different shape, a signed char handed
to <cctype> where a negative value is undefined; every such call here
(chessboard.cpp's toupper, tolower and isupper) already casts correctly. So
the check is NARROWED to ignore the qint8 typedef rather than switched off,
and both directions were verified: the qint8 readers go silent, and the same
field declared `signed char` is still reported.

Two things about that option worth knowing. It is spelled CharTypedefsToIgnore
-- an older name for it is missing an `e` and is silently ignored rather
than refused -- and bugprone-signed-char-misuse and cert-str34-c are aliases
that each carry their own copy, so setting one leaves the other reporting.
Both came from `clang-tidy --dump-config` rather than from memory.

Misplaced widening casts, all five: an index computed in int and widened
afterwards, which does not help if the int arithmetic overflowed. Each now
widens first. None could overflow at these board sizes; the point is that
the code no longer relies on that.

Throwing static initialisation, all eight. Seven are a file-scope QString
built from QStringLiteral, and they become functions returning it -- the
shape scores.h already uses for its own keys, so this is the project's own
idiom rather than a new one. QStringLiteral's data is static, so the
returned copy costs nothing, and the key strings are byte-identical, so no
stored setting moves. The eighth is a static mt19937, now function-local.

Three value parameters in DraughtsBoard::collectJumps, which is the
recursive capture search: it copied two vectors and a whole board at every
step and then copied them again into the locals it actually mutates. Now
const references.

Three implicit widenings of a multiplication, one of them introduced by this
session's own cache-byte bound.

Verified: run-clang-tidy over src/*.cpp reports nothing, and exits 0 with
WarningsAsErrors set. ctest 9/9. The Canasta AI ladder is unchanged to the
digit -- 22/24 +2538, 22/24 +3547, 69/120 +424, 121/240 -1.

Not done, and said rather than left to be discovered: nothing in CI runs
clang-tidy, so the enforcement binds a manual run and nothing else. Wiring
it into ci.yml is a separate decision -- it costs runner time on every push
and scripts/local-ci.sh would then run it on every local gate too.

+++ src/hubwindow.cpp
@@ -342,7 +342,7 @@ void snakeTile(QPainter& p, const QRectF& r)
     p.setPen(Qt::NoPen);
     const QPoint body[5] = { { 1, 3 }, { 1, 2 }, { 2, 2 }, { 3, 2 }, { 3, 1 } };
     for (int i = 0; i < 5; ++i) {
-        p.setBrush(QColor::fromHsvF(0.33, 0.55, 0.85 - i * 0.07));
+        p.setBrush(QColor::fromHsvF(0.33f, 0.55f, float(0.85 - i * 0.07)));
         p.drawRoundedRect(QRectF(r.left() + body[i].x() * cell + cell * 0.08,
                                  r.top() + body[i].y() * cell + cell * 0.08,
                                  cell * 0.84, cell * 0.84),
